// frontend/src/types/Task.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskConfig } from "./TaskConfig";
import type { TaskStatus } from "./TaskStatus";
import type { TaskType } from "./TaskType";
import type { JsonValue } from "./serde_json/JsonValue";

export type Task = { id: string, task_type: TaskType, status: TaskStatus, progress: number, config: TaskConfig, result?: JsonValue, error_message?: string, created_at: number, started_at?: number, completed_at?: number, updated_at: number, };