// frontend/src/types/TaskEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Task } from "./Task";
import type { TaskStatus } from "./TaskStatus";
import type { JsonValue } from "./serde_json/JsonValue";

export type TaskEvent = { "type": "create", task: Task, } | { "type": "progress", task_id: string, progress: number, status: TaskStatus, } | { "type": "status", task_id: string, status: TaskStatus, } | { "type": "complete", task_id: string, result?: JsonValue, } | { "type": "fail", task_id: string, error: string, };